DTMESH-407 :- Implement the support of STA association and EAPOL handling (#246)
Reason for change:-
EAPOL_OVER_NL support should be enabled for station interfaces.
Bring up DT follower mode device

Testing Performed:-
Verified DT follower device comes up online in the NOC and able to
associate bhaul or private clients to the device
Verified XLE device comes up online after the changes and
able to verify basic sanity

+++ src/wifi_hal_nl80211_utils.c
@@ -1798,7 +1798,7 @@ BOOL get_ie_ext_by_eid(unsigned int eid, unsigned char *buff, unsigned int buff_
     return false;
 }
 
-BOOL get_ie_by_eid(unsigned int eid, unsigned char *buff, unsigned int buff_len, unsigned char **ie_out, unsigned short *ie_out_len)
+BOOL get_ie_by_eid(unsigned int eid, unsigned char *buff, unsigned int buff_len, unsigned char **ie_out, size_t *ie_out_len)
 {
     ieee80211_tlv_t *ie = NULL;
     signed int len;
